requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
dj-email-url==1.0.6
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-cache-url==3.4.5
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-spectacular==0.27.2
environs==11.0.0
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
marshmallow==3.21.3
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
